Fix date pill selection resetting to today on app foreground (#100)
* Fix date pill selection resetting to today on app foreground

SyncViewModel.onAppForeground() ran on every ON_RESUME and reset the
selected date whenever it differed from today, even with no actual
midnight rollover. Picking a past date via the dashboard date pill and
simply backgrounding/foregrounding the app immediately snapped it back
to today.

SelectedDateRepository now tracks lastKnownToday and exposes
advanceTodayIfNeeded(), which only advances the selection when the user
was passively viewing today and the calendar day genuinely rolled over,
leaving any explicit historical pick untouched.


* Fix review feedback on date-pill foreground fix (PR #100)

- Sequence advanceTodayIfNeeded() before the permission-check/sync flow
  in a single foregroundCheckJob coroutine, removing the race where
  evaluateAndSync() could read the date before it advanced.
- advanceTodayIfNeeded() now also corrects the selected date if the
  clock moves backward (or timezone travel) leaves it in the future
  relative to the recomputed "today".
- updateSelectedDate/selectPreviousDay/selectNextDay now refresh
  lastKnownToday so an explicit pick made after a missed midnight
  rollover isn't mistaken for a stale "today" on the next foreground.
- Add a regression test asserting advanceTodayIfNeeded() runs before
  evaluateAndSync() when foregrounding with permissions granted.


* fix failing tests

---------

+++ app/src/test/kotlin/app/readylytics/health/domain/scoring/ScoringPointInTimeRegressionTest.kt
@@ -406,7 +406,7 @@ class ScoringPointInTimeRegressionTest {
 
                 assertEquals(
                     targetDate
-                        .minusDays((ScoringConstants.CHRONIC_DAYS * 2).toLong())
+                        .minusDays(ScoringConstants.CHRONIC_DAYS * 2)
                         .atStartOfDay(zoneId)
                         .toInstant()
                         .toEpochMilli(),
